debug

+++ portiloop_detector/pareto_search.py
@@ -383,7 +383,7 @@ class SurrogateModel(nn.Module):
         self.device = device
 
     def forward(self, config_dict):
-        x_list = [config_dict["seq_len"],
+        x_list = [float(config_dict["seq_len"]),
                   config_dict["nb_channel"],
                   # config_dict["dropout"],
                   config_dict["hidden_size"],
@@ -400,7 +400,7 @@ class SurrogateModel(nn.Module):
                   config_dict["dilation_conv"],
                   config_dict["dilation_pool"]]
 
-        x_tensor = torch.tensor(x_list, dtype=torch.double).to(self.device)
+        x_tensor = torch.tensor(x_list).to(self.device)
 
         x_tensor = F.relu(self.d1(self.fc1(x_tensor)))
         x_tensor = F.relu(self.d2(self.fc2(x_tensor)))
